Maintain a single generated password across page views
