--- install/google_cloud_sql/_01_create_monitoring_user.mdx
@@ -1,46 +1,0 @@
-import MonitoringUserSetupInstructions from "../../components/MonitoringUserSetupInstructions";
-
-export const VerifyConnection = ({password}) => {
-  return (
-    <CodeBlock>PGPASSWORD={password} psql -h 1.2.3.4 -d mydatabase -U pganalyze</CodeBlock>
-  )
-}
-
-We recommend creating a separate monitoring user on your PostgreSQL database for pganalyze.
-
-## Enabling pg_stat_statements
-
-Connect to your database with **cloudsqlsuperuser / alloydbsuperuser** privileges (i.e. the initial user that was created on instance creation), e.g. using `psql`.
-
-Run the following SQL commands to enable the `pg_stat_statements` extension, and make sure it was installed correctly:
-
-```
-CREATE EXTENSION IF NOT EXISTS pg_stat_statements;
-SELECT * FROM pg_stat_statements LIMIT 1;
-```
-
-## Create monitoring user
-
-Connect to the database you will be monitoring, again as a user with **cloudsqlsuperuser / alloydbsuperuser** privileges.
-Then run the following to create a monitoring user (we've generated a random password for you, but you
-can replace it with one of your choosing):
-
-<MonitoringUserSetupInstructions password={props.password} />
-
-Note that it is important you run these as a user with **cloudsqlsuperuser / alloydbsuperuser** privileges in order to pass down the full access to statistics tables.
-
-## Verify connection
-
-Let's make sure we can connect to the database with the new user:
-
-<VerifyConnection password={props.password} />
-
-Write down the username and password of the monitoring user, we will use it in the last step of this tutorial.
-
----
-
-Next we continue by installing the pganalyze collector:
-
-<Link className="btn btn-success" to="02_install_the_collector">
-  Proceed to Step 2: Install the collector
-</Link>
